+++ test/runtests.jl
@@ -42,6 +42,7 @@ scs_solver = SCSSolver(max_iters=500000, acceleration_lookback=1, verbose=0)
 include("common.jl")
 
 @testset "PowerModels" begin
+
     include("matpower.jl")
 
     include("pti.jl")
@@ -80,6 +81,5 @@ include("common.jl")
 
     include("warmstart.jl")
 
-    #include("docs.jl")
-
+    include("docs.jl")
 end